move common github code out
