+++ packages/branch-protector/package.json
@@ -10,12 +10,10 @@
 	},
 	"author": "guardian",
 	"dependencies": {
-		"@aws-sdk/client-secrets-manager": "^3.441.0",
 		"@aws-sdk/client-sqs": "^3.438.0",
 		"@aws-sdk/client-ssm": "^3.438.0",
 		"@aws-sdk/types": "^3.433.0",
 		"@guardian/anghammarad": "^1.8.1",
-		"@octokit/auth-app": "^6.0.1",
 		"@octokit/types": "^12.0.0",
 		"octokit": "^3.1.1"
 	}
