+++ packages/branch-protector/src/config.ts
@@ -1,14 +1,5 @@
-import { SecretsManager } from '@aws-sdk/client-secrets-manager';
-import { type StrategyOptions } from '@octokit/auth-app';
-
-//TODO: move to a common place
-export function getEnvOrThrow(key: string): string {
-	const value: string | undefined = process.env[key];
-	if (value === undefined) {
-		throw new Error(`Environment variable ${key} is not set.`);
-	}
-	return value;
-}
+import { getEnvOrThrow, getGitHubAppConfig } from 'common/functions';
+import type { GitHubAppConfig } from 'common/types';
 
 export interface Config {
 	/**
@@ -24,10 +15,7 @@ export interface Config {
 	/**
 	 * Auth configuration for the GitHub App.
 	 */
-	githubAppConfig: {
-		strategyOptions: StrategyOptions;
-		installationId: string;
-	};
+	githubAppConfig: GitHubAppConfig;
 
 	/**
 	 * SNS topic to use for Anghammarad.
@@ -40,37 +28,11 @@ export interface Config {
 	queueUrl: string;
 }
 
-interface GithubAppSecret {
-	appId: string;
-	base64PrivateKey: string;
-	clientId: string;
-	clientSecret: string;
-	installationId: string;
-}
-
-async function getGithubAppSecretJson(): Promise<GithubAppSecret> {
-	const secretsManager = new SecretsManager();
-
-	const secret = await secretsManager.getSecretValue({
-		SecretId: getEnvOrThrow('GITHUB_APP_SECRET'),
-	});
-
-	const secretJson = JSON.parse(secret.SecretString ?? '{}') as GithubAppSecret;
-	return secretJson;
-}
-
 export async function getConfig() {
-	const secretJson = await getGithubAppSecretJson();
 	const config: Config = {
 		app: getEnvOrThrow('APP'),
 		stage: process.env['STAGE'] ?? 'DEV',
-		githubAppConfig: {
-			strategyOptions: {
-				...secretJson,
-				privateKey: atob(secretJson.base64PrivateKey),
-			},
-			installationId: secretJson.installationId,
-		},
+		githubAppConfig: await getGitHubAppConfig(),
 		anghammaradSnsTopic: getEnvOrThrow('ANGHAMMARAD_SNS_ARN'),
 		queueUrl: getEnvOrThrow('BRANCH_PROTECTOR_QUEUE_URL'),
 	};
