+++ packages/common/package.json
@@ -6,7 +6,9 @@
 		"test": "jest --detectOpenHandles --config ../../jest.config.js --selectProjects common"
 	},
 	"dependencies": {
+		"@aws-sdk/client-secrets-manager": "^3.441.0",
 		"@guardian/anghammarad": "^1.8.1",
-		"@aws-sdk/credential-providers": "^3.444.0"
+		"@aws-sdk/credential-providers": "^3.444.0",
+		"@octokit/auth-app": "^6.0.1"
 	}
 }
