+++ packages/common/src/types.ts
@@ -1,4 +1,19 @@
+import { type StrategyOptions } from '@octokit/auth-app';
+
+export type GithubAppSecret = {
+	appId: string;
+	base64PrivateKey: string;
+	clientId: string;
+	clientSecret: string;
+	installationId: string;
+};
+
 export interface UpdateBranchProtectionEvent {
 	fullName: string; // in the format of owner/repo-name
 	teamNameSlugs: string[];
 }
+
+export type GitHubAppConfig = {
+	strategyOptions: StrategyOptions;
+	installationId: string;
+};
